+++ src/main/java/com/uade/tpo/demo/controllers/products/ProductsRequest.java
@@ -1,4 +1,6 @@
 package com.uade.tpo.demo.controllers.products;
+import java.math.BigDecimal;
+
 import lombok.Data;
 
 @Data
@@ -7,8 +9,8 @@ public class ProductsRequest {
     private String name;
     private String description;
     private Long categoryId; // <-- solo el ID, luego busco dentro del controller a la categoria
-    private Double price;
+    private BigDecimal price;
     private Integer stock;
     private String imageUrl;
-    private Double discount;
+    private BigDecimal discount;
 }
